cst-azureml-llm-deployable-v2/infra/requirements.deploy.txt
azure-ai-ml==1.31.0
azure-identity==1.25.1
python-dotenv==1.0.1
